backend-flask/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.2.0
cloudinary==1.40.0
gunicorn==21.2.0
